fix signed/unsigned comparison warning
git-svn-id: svn://scribus.net/trunk/Scribus@22684 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/vgradient.cpp
@@ -113,7 +113,7 @@ VGradient& VGradient::operator=( const VGradient& gradient )
 
 bool VGradient::operator==(const VGradient &gradient) const 
 {
-	if (static_cast<uint>(m_colorStops.count()) != gradient.Stops())
+	if (m_colorStops.count() != gradient.Stops())
 		return false;
 	const QList<VColorStop*>& cs = gradient.colorStops();
 	bool retVal = true;
